Move parseAndValidateFiles into the validator code
